feat: use MapContextProvider to display controls on the map MapContextProvider passes the context of the map to its children such as MapControls so that geolocation and navigation can function on the map canvas.
