+++ app/map/page.tsx
@@ -1,9 +1,9 @@
-import MapContainer from "../_components/Map";
+import MapContainer from "../_components/MapContainer";
 
 export default function Home(): JSX.Element {
   return (
-    <main>
-      <MapContainer />
-    </main>
+      <div className="absolute overflow-hidden inset-0 bg-mapBg">
+        <MapContainer />
+      </div>
   );
 }
